[fix] some bugs during preintegration and optimization; And fixed a bug during erase chain when sliding window is full

+++ state_estimation/include/g2oTypes/Edges.h
@@ -87,7 +87,7 @@ class EdgePriorGyrBias : public g2o::BaseUnaryEdge<3, Eigen::Vector3d, VertexGyr
 public:
     EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 
-    EdgePriorGyrBias(const Eigen::Vector3d &bprior_) : bprior(bprior_.cast<double>()) {}
+    EdgePriorGyrBias(const Eigen::Vector3d &bprior_) : bprior(bprior_) {}
 
     virtual bool read(std::istream &is) { return false; }
     virtual bool write(std::ostream &os) const { return false; }
